% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fontHind-package.r
\docType{package}
\name{fontHind}
\alias{fontHind}
\alias{fontHind-package}
\title{Additional ggplot2 themes using Hind fonts.}
\description{
This is an add-on pacakge for \code{\link{hrbrthemes}} pacakge.
 It provides seven ggplot2 themes based on
  \href{https://fonts.google.com/specimen/Hind}{Hind} fonts.
}
\section{Hind Fonts}{

Hind is an Open Source typeface supporting the Devanagari and Latin scripts.
Developed explicitly for use in User Interface design, the Hind font family includes five styles.
More information about the font can be found at \url{https://fonts.google.com/specimen/Hind}.
}

\section{One-time setup}{


Before using this theme you need to call \code{\link{import_hind}()} which will
 import the Hind fonts in your \code{\link{extrafont}} database.
 Then you will need to install the Hind fonts in the directory mentioned in the output
 of \code{\link{import_hind}()} as per your operating system's way of installing fonts.
}

