% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hind.R
\name{import_hind}
\alias{import_hind}
\title{Import Hind font for use in \code{\link[hrbrthemes]{hrbrthemes-package}}.}
\usage{
import_hind()
}
\description{
Import Hind font for use in \code{\link[hrbrthemes]{hrbrthemes-package}}.
}
\note{
This will take care of ensuring PDF/PostScript usage. The location of the
  font directory is displayed after the base import is complete. It is highly
  recommended that you install them on your system the same way you would any
  other font you wish to use in other programs.
}
